tests/submission/Microsoft/forms/datalistelement.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>datalist element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="datalist element" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>
        <div id=log></div>
    
        <datalist style="display: none" id="datalist1">
          <option>1</option>
          <option>2</option>
          <option>3</option>
          <option>4</option>
        </datalist>

        <script type="text/javascript">

        test(function() {assert_true(HTMLDataListElement && document.getElementsByTagName("datalist")[0] instanceof HTMLDataListElement)}, "HTMLDataListElement support");
        test(function() {assert_equals(document.getElementsByTagName("datalist")[0].options.item(0), document.getElementsByTagName("option")[0])}, "HTMLDataListElement options item collection support");
        </script>
    </body>
</html>
